requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
python-dotenv==1.1.0
anthropic==0.89.0
pymupdf==1.25.5
python-multipart==0.0.20
